+++ lib/experimental/Information/Communities/Celebration/types.ts
@@ -1,0 +1,31 @@
+export type CelebrationProps = {
+  firstName: string
+  lastName: string
+  src?: string
+  type?: "birthday" | "anniversary" | "first-day"
+  typeLabel: string
+  date: {
+    day: number
+    month: string
+  }
+}
+
+export const EMOJI_MAP = {
+  birthday: "🎂",
+  anniversary: "🎉",
+  "first-day": "💼",
+} as const
+
+export const BACKGROUND_COLORS = {
+  viridian: "bg-[hsl(theme(colors.viridian.50)_/_0.2)]",
+  malibu: "bg-[hsl(theme(colors.malibu.50)_/_0.2)]",
+  yellow: "bg-[hsl(theme(colors.yellow.50)_/_0.2)]",
+  purple: "bg-[hsl(theme(colors.purple.50)_/_0.2)]",
+  lilac: "bg-[hsl(theme(colors.lilac.50)_/_0.2)]",
+  barbie: "bg-[hsl(theme(colors.barbie.50)_/_0.2)]",
+  smoke: "bg-[hsl(theme(colors.smoke.50)_/_0.2)]",
+  army: "bg-[hsl(theme(colors.army.50)_/_0.2)]",
+  flubber: "bg-[hsl(theme(colors.flubber.50)_/_0.2)]",
+  indigo: "bg-[hsl(theme(colors.indigo.50)_/_0.2)]",
+  camel: "bg-[hsl(theme(colors.camel.50)_/_0.2)]",
+} as const
